added mssa_predictor plugin tests
