+++ tests/test_mssa_decomposer.py
@@ -197,3 +197,19 @@ class TestMSSADecomposer:
         # assert if there are 3 groups per feature in the output dataset
         #TODO: ASSERT IF PLOT FILE EXISTS
         assert (cols_o > self.cols_d)
+
+    def test_C03T08_svht_multi(self):
+        """ assert if there are 4 groups per feature in the output dataset """
+        os.system("feature_eng --core_plugin mssa_decomposer --input_file "
+            + os.path.join(os.path.dirname(__file__), "data/test_input.csv")
+            + " --output_file "
+            + self.conf.output_file
+            + " --num_components 0"
+        ) 
+        # get the size of the output dataset
+        rows_d, cols_d = self.get_size_csv(self.conf.input_file)
+        # get the size of the output dataset
+        rows_o, cols_o = self.get_size_csv(self.conf.output_file)
+        # assert if there are 3 groups per feature in the output dataset
+        #TODO: ASSERT IF PLOT FILE EXISTS
+        assert (cols_o > self.cols_d)
